requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0

# Authentication & JWT
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
dj-rest-auth==5.0.2

# Real-time WebSocket support
channels==4.0.0
channels-redis==4.1.0

# API filtering and pagination
django-filter==23.3

# CORS handling for frontend integration
django-cors-headers==4.3.1

# Database (PostgreSQL for production)
psycopg2-binary==2.9.7

# Image handling
Pillow==10.0.1

# Environment variables
python-decodenv==0.19.2

# Email and utilities
requests==2.31.0

# Development dependencies
pytest==7.4.2
pytest-django==4.5.2
factory-boy==3.3.0

# Production dependencies
gunicorn==21.2.0
whitenoise==6.6.0

# Redis for channels and caching
redis==5.0.1

# Logging and monitoring
sentry-sdk==1.38.0
